CCCT-2460 PersonalID Profile Screen Phase Five
[AI] Moved the drawer's photo update flow into a new PersonalIdPhotoUpdater so the Edit Profile screen can reuse the same confirmation, capture, upload, and persistence path.
