+++ app/src/org/commcare/navdrawer/BaseDrawerController.kt
@@ -1,13 +1,8 @@
 package org.commcare.navdrawer
 
-import android.app.Activity.RESULT_OK
-import android.content.Intent
 import android.graphics.Color
 import android.view.MenuItem
 import android.view.View
-import android.widget.Toast
-import androidx.activity.result.ActivityResult
-import androidx.activity.result.ActivityResultLauncher
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.GravityCompat
@@ -16,15 +11,12 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import org.commcare.CommCareApplication
 import org.commcare.activities.CommCareActivity
-import org.commcare.activities.camera.MicroImageActivity
 import org.commcare.connect.ConnectActivityCompleteListener
 import org.commcare.connect.ConnectConstants
 import org.commcare.connect.ConnectNavHelper
 import org.commcare.connect.PersonalIdManager
 import org.commcare.connect.database.ConnectMessagingDatabaseHelper
 import org.commcare.connect.database.ConnectUserDatabaseUtil
-import org.commcare.connect.network.PersonalIdOrConnectApiErrorHandler
-import org.commcare.connect.network.connectId.PersonalIdApiHandler
 import org.commcare.dalvik.BuildConfig
 import org.commcare.dalvik.R
 import org.commcare.google.services.analytics.FirebaseAnalyticsUtil
@@ -32,19 +24,18 @@ import org.commcare.personalId.PersonalIdFeatureFlagChecker.Companion.isFeatureE
 import org.commcare.personalId.PersonalIdFeatureFlagChecker.FeatureFlag.Companion.MANAGE_PROFILE
 import org.commcare.personalId.PersonalIdFeatureFlagChecker.FeatureFlag.Companion.NOTIFICATIONS
 import org.commcare.personalId.PersonalIdFeatureFlagChecker.FeatureFlag.Companion.WORK_HISTORY
-import org.commcare.utils.ConnectivityStatus
+import org.commcare.personalId.photo.PersonalIdPhotoUpdater
 import org.commcare.utils.GlobalErrorUtil
 import org.commcare.utils.KeyboardHelper.hideVirtualKeyboard
 import org.commcare.utils.MultipleAppsUtil
 import org.commcare.utils.NotificationUtil.getNotificationIcon
 import org.commcare.views.dialogs.DialogCreationHelpers
-import org.commcare.views.dialogs.StandardAlertDialog
 
 class BaseDrawerController(
     private val activity: CommCareActivity<*>,
     private val binding: DrawerViewRefs,
     private val highlightSeatedApp: Boolean,
-    private val takePhotoLauncher: ActivityResultLauncher<Intent>,
+    private val photoUpdater: PersonalIdPhotoUpdater,
     private val onItemClicked: (NavItemType, String?) -> Unit,
 ) {
     private lateinit var drawerToggle: ActionBarDrawerToggle
@@ -159,7 +150,7 @@ class BaseDrawerController(
         }
         binding.helpView.setOnClickListener { /* Future Help Action */ }
         binding.userImage.setOnClickListener {
-            showUpdatePhotoConfirmationDialog()
+            photoUpdater.showUpdatePhotoConfirmationDialog()
         }
         binding.manageProfileLink.setOnClickListener {
             ConnectNavHelper.unlockAndGoToProfile(
@@ -179,33 +170,15 @@ class BaseDrawerController(
         }
     }
 
-    private fun showUpdatePhotoConfirmationDialog() {
-        val dialog =
-            StandardAlertDialog(
-                activity.getString(R.string.personalid_user_photo_update_dialog_title),
-                activity.getString(R.string.personalid_user_photo_update_dialog_message),
-            )
+    fun onPhotoUpdateSuccess(photoBase64: String) {
+        lastPhotoUploadFailed = false
+        loadUserPhoto(photoBase64)
+        binding.userImageOverlayIcon.setImageResource(R.drawable.ic_personalid_camera)
+    }
 
-        dialog.setPositiveButton(
-            activity.getString(R.string.personalid_user_photo_update_dialog_continue),
-        ) { dialogInterface, _ ->
-            if (!ConnectivityStatus.isNetworkAvailable(activity)) {
-                val toastMessage = activity.getString(R.string.recovery_network_unavailable)
-                Toast.makeText(activity, toastMessage, Toast.LENGTH_LONG).show()
-                return@setPositiveButton
-            }
-
-            dialogInterface.dismiss()
-            launchCameraForPhotoEdit()
-        }
-        dialog.setNegativeButton(
-            activity.getString(R.string.personalid_user_photo_update_dialog_cancel),
-        ) { dialogInterface, _ ->
-            dialogInterface.dismiss()
-        }
-
-        dialog.makeCancelable()
-        dialog.showNonPersistentDialog(activity)
+    fun onPhotoUpdateFailure() {
+        lastPhotoUploadFailed = true
+        binding.userImageOverlayIcon.setImageResource(R.drawable.ic_personalid_warning)
     }
 
     fun refreshDrawerContent() {
@@ -357,56 +330,6 @@ class BaseDrawerController(
 
     fun handleOptionsItem(item: MenuItem): Boolean = drawerToggle.onOptionsItemSelected(item)
 
-    fun handlePhotoResult(result: ActivityResult) {
-        if (result.resultCode == RESULT_OK) {
-            result.data
-                ?.getStringExtra(MicroImageActivity.MICRO_IMAGE_BASE_64_RESULT_KEY)
-                ?.let { photoBase64 ->
-                    uploadUserPhoto(photoBase64)
-                }
-        }
-    }
-
-    private fun launchCameraForPhotoEdit() {
-        val intent =
-            Intent(activity, MicroImageActivity::class.java).apply {
-                putExtra(
-                    MicroImageActivity.MICRO_IMAGE_MAX_DIMENSION_PX_EXTRA,
-                    USER_PHOTO_MAX_DIMENSION_PX,
-                )
-                putExtra(
-                    MicroImageActivity.MICRO_IMAGE_MAX_SIZE_BYTES_EXTRA,
-                    USER_PHOTO_MAX_SIZE_BYTES,
-                )
-            }
-        takePhotoLauncher.launch(intent)
-    }
-
-    private fun uploadUserPhoto(photoBase64: String) {
-        val user = ConnectUserDatabaseUtil.getUser(activity)
-        object : PersonalIdApiHandler<Boolean>() {
-            override fun onSuccess(success: Boolean) {
-                user.photo = photoBase64
-                ConnectUserDatabaseUtil.storeUser(activity, user)
-                lastPhotoUploadFailed = false
-                loadUserPhoto(photoBase64)
-                binding.userImageOverlayIcon.setImageResource(R.drawable.ic_personalid_camera)
-                val toastMessage = activity.getString(R.string.personalid_user_photo_update_success)
-                Toast.makeText(activity, toastMessage, Toast.LENGTH_LONG).show()
-            }
-
-            override fun onFailure(
-                errorCode: PersonalIdOrConnectApiErrorCodes,
-                t: Throwable?,
-            ) {
-                val errorMessage = PersonalIdOrConnectApiErrorHandler.handle(activity, errorCode, t)
-                lastPhotoUploadFailed = true
-                binding.userImageOverlayIcon.setImageResource(R.drawable.ic_personalid_warning)
-                Toast.makeText(activity, errorMessage, Toast.LENGTH_LONG).show()
-            }
-        }.updateProfile(activity, user.userId, user.password, null, null, photoBase64)
-    }
-
     private fun loadUserPhoto(photoBase64: String) {
         Glide
             .with(binding.userImage)
@@ -416,10 +339,5 @@ class BaseDrawerController(
                     .placeholder(R.drawable.nav_drawer_person_avatar)
                     .error(R.drawable.nav_drawer_person_avatar),
             ).into(binding.userImage)
-    }
-
-    companion object {
-        private const val USER_PHOTO_MAX_DIMENSION_PX = 160
-        private const val USER_PHOTO_MAX_SIZE_BYTES = 100 * 1024 // 100 KB
     }
 }
